+++ BookStoreEJB/ejbModule/ejb/remote/BookManager.java
@@ -12,7 +12,7 @@ public interface BookManager {
 	public void deleteBook(Book book);
 	public Book updateBookInfo(Book book);
 	
-	public Book getBookById (int bookId);
+	public Book getBookById (long bookId);
 	
 	public List<Book> searchBooksByTitle(String title);
 	public List<Book> searchBooksByAuthor(String author);
